// generated/src/app/api-models/stage-era-company.model.ts
/* Copyright (c) 2020 . All Rights Reserved. */

export class StageEraCompany {
  batchId: string;
  checkAmt: number;
  abaRoutingNumber: string;
  bankAccountNum: string;
  companyCode: string;
  vendAbaRoutingNumber: string;
  vendBankAccountNumber: string;
  checkDate: Date;
  checkOrEftNumber: string;
  payerIrsTaxId: string;
  payerName: string;
  payerAddressLine1: string;
  payerAddressLine2: string;
  payerCity: string;
  payerState: string;
  payerZipCode: string;
  payeeFullName: string;
  payeeIrsTaxId: string;
  securityCode: string;
  insertDatetime: Date;
  insertUser: string;
  insertProcess: string;
  updateDatetime: Date;
  updateUser: string;
  updateProcess: string;
}